+++ README.md
@@ -3,8 +3,6 @@
 [![Hex.pm](https://img.shields.io/hexpm/dt/closure_table.svg?maxAge=2592000)](https://hex.pm/packages/closure_table)
 [![Hexdocs.pm](https://img.shields.io/badge/api-hexdocs-brightgreen.svg)](https://hexdocs.pm/closure_table)
 
-> this library provides two adapters: an in-memory one, and one for using the closure-table solution with Ecto; for your testing and development convenience.
-
 The Closure Table solution is a simple and elegant way of storing hierarchies. It involves storing all paths through a tree, not just those with a direct parent-child relationship. You may want to chose this model, over the [Nested Sets model](https://en.wikipedia.org/wiki/Nested_set_model), should you need referential integrity and to assign nodes to multiple trees.
 
 Throughout the various examples and tests, we will refer to the hierarchies depicted below, where we're modeling a hypothetical forum-like discussion between [Rolie, Olie and Polie](https://www.youtube.com/watch?v=LTkmaE_QWMQ), and their debate around the usefulness of this implementation :)
@@ -13,39 +11,9 @@ Throughout the various examples and tests, we will refer to the hierarchies depi
 
 ## Quick start
 
-To start, you can simply use one `Adapter` from the ones provided, same way you'd use the Ecto's own Repo:
+### TODO
 
-```elixir
-defmodule CTT do
-  use CTE,
-    otp_app: :cte,
-    adapter: CTE.Adapter.Memory,
-    nodes: %{
-      1 => %{id: 1, author: "Olie", comment: "Is Closure Table better than the Nested Sets?"},
-      2 => %{id: 2, author: "Rolie", comment: "It depends. Do you need referential integrity?"},
-      3 => %{id: 3, author: "Olie", comment: "Yeah."}
-    },
-    paths: [[1, 1, 0], [1, 2, 1], [1, 3, 2], [2, 2, 0], [2, 3, 1], [3, 3, 0]]
-end
-```
-
-With the configuration above, the `:nodes` attribute is a map containing the comments our interlocutors made; these are "nodes", in CTE's parlance. When using the `CTE.Adapter.Ecto` implementation, the `:nodes` attribute will be a Schema (or a table name! In our initial implementation, the nodes definitions must have at least the `:id`, as one of their properties. This caveat will be lifted in a later implementation. The `:paths` attribute represents the parent-child relationship between the comments.
-
-Add the `CTM` module to your main supervision tree:
-
-```elixir
-defmodule CTM.Application do
-  @moduledoc false
-
-  use Application
-
-  def start(_type, _args) do
-    opts = [strategy: :one_for_one, name: CTM.Supervisor]
-
-    Supervisor.start_link([CTM], opts)
-  end
-end
-```
+````
 
 And then using `iex -S mix`, for quickly experimenting with the CTE API, let's find the descendants of comment #1:
 
@@ -62,7 +30,7 @@ Is Closure Table better than the Nested Sets?
 └── It depends. Do you need referential integrity?
    └── Yeah.
 
-```
+````
 
 Please check the docs for more details and return from more updates!
 
@@ -82,7 +50,7 @@ by adding `closure_table` to your list of dependencies in `mix.exs`:
 ```elixir
 def deps do
   [
-    {:closure_table, "~> 1.1"}
+    {:closure_table, "~> 2.0"}
   ]
 end
 ```
