+++ examples/ct_ecto/config/dev.exs
@@ -1,7 +1,7 @@
 import Config
 
 config :ct, CT.Repo,
-  database: "ct_ecto_dev",
+  database: "ct_dev",
   username: "postgres",
   password: "postgres",
   hostname: "localhost",
