+++ examples/ct_ecto/lib/ct/my_cte.ex
@@ -3,8 +3,6 @@ defmodule CT.MyCTE do
   Comments hierarchy
   """
   use CTE,
-    otp_app: :ct,
-    adapter: CTE.Adapter.Ecto,
     repo: CT.Repo,
     nodes: CT.Comment,
     paths: CT.TreePath
