--- examples/ct_memory/README.md
@@ -1,20 +1,0 @@
-# CTM
-
-**TODO: Add description**
-
-## Installation
-
-If [available in Hex](https://hex.pm/docs/publish), the package can be installed
-by adding `ct_memory` to your list of dependencies in `mix.exs`:
-
-```elixir
-def deps do
-  [
-    {:ct_memory, "~> 0.1.1"}
-  ]
-end
-```
-
-Documentation can be generated with [ExDoc](https://github.com/elixir-lang/ex_doc)
-and published on [HexDocs](https://hexdocs.pm). Once published, the docs can
-be found at [https://hexdocs.pm/ct_memory](https://hexdocs.pm/ct_memory).
