--- examples/ct_memory/lib/ctm/application.ex
@@ -1,11 +1,0 @@
-defmodule CTM.Application do
-  @moduledoc false
-
-  use Application
-
-  def start(_type, _args) do
-    opts = [strategy: :one_for_one, name: CTM.Supervisor]
-
-    Supervisor.start_link([CTM], opts)
-  end
-end
