--- examples/ct_memory/test/test_helper.exs
@@ -1,1 +1,0 @@
-ExUnit.start()
