+++ test/cte_test.exs
@@ -43,16 +43,12 @@ defmodule CTE.Ecto.Test do
     Comments hierarchy
     """
     use CTE,
-      otp_app: :cte,
-      adapter: CTE.Adapter.Ecto,
       repo: Repo,
       nodes: Comment,
       paths: TreePath
   end
 
   setup_all do
-    start_supervised!({CH, []})
-
     Repo.delete_all(Comment)
     Repo.delete_all(Author)
     Repo.delete_all(TreePath)
@@ -184,7 +180,7 @@ defmodule CTE.Ecto.Test do
       assert {:ok, [%Comment{text: "w⦿‿⦿t!"}]} =
                CH.descendants(9, limit: 1, itself: true, nodes: true)
 
-      assert :ok == CH.delete(9, limit: 1)
+      assert {:ok, %{deleted: 4, updated: 0}} = CH.delete(9, limit: 1)
 
       assert {:ok, []} == CH.descendants(9, limit: 1, itself: true, nodes: true)
       assert {:ok, []} == CH.descendants(9, limit: 1)
@@ -192,19 +188,19 @@ defmodule CTE.Ecto.Test do
 
     test "delete subtree; comment #6 and its descendants" do
       assert {:ok, [6, 8, 9]} == CH.descendants(6, itself: true)
-      assert :ok == CH.delete(6, limit: 0)
+      assert {:ok, %{deleted: 11, updated: 0}} == CH.delete(6, limit: 0)
       assert {:ok, []} == CH.descendants(6, itself: true)
     end
 
     test "delete subtree w/o any leafs; comment #5 and its descendants" do
       assert {:ok, [5]} == CH.descendants(5, itself: true)
-      assert :ok == CH.delete(5)
+      assert {:ok, %{deleted: 3, updated: 0}} == CH.delete(5)
       assert {:ok, []} == CH.descendants(5, itself: true)
     end
 
     test "delete whole tree, from its root; comment #1" do
       assert {:ok, [1, 2, 4, 3, 5, 6, 7, 8, 9]} == CH.descendants(1, itself: true)
-      assert :ok == CH.delete(1, limit: 0)
+      assert {:ok, %{deleted: 26, updated: 0}} == CH.delete(1, limit: 0)
       assert {:ok, []} == CH.descendants(1, itself: true)
     end
 
@@ -224,7 +220,7 @@ defmodule CTE.Ecto.Test do
       assert {:ok, [1, 4]} == CH.ancestors(6)
       assert {:ok, [8, 9]} == CH.descendants(6)
 
-      assert {:ok, {9, nil}} = CH.move(6, 3)
+      assert {:ok, %{inserted: 9, deleted: 6}} = CH.move(6, 3)
 
       assert {:ok, list} = CH.ancestors(6)
       assert MapSet.subset?(MapSet.new([1, 2, 3]), MapSet.new(list))
@@ -320,7 +316,7 @@ defmodule CTE.Ecto.Test do
 
       print_io = fn ->
         # limit: 1
-        assert :ok == CH.delete(3)
+        assert {:ok, %{deleted: 4, updated: 2}} == CH.delete(3)
         {:ok, tree} = CH.tree(1)
         CTE.Utils.print_tree(tree, 1, callback: &{&1, "(#{&1}) #{&2[&1].text}"})
       end
